pyproject.toml
[project]
name = "planner"
version = "0.1.0"
description = "Confidently navigate LLM deployments from concept to production"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    # LLM integration
    "ollama==0.3.3",
    # Data handling
    "pandas==2.2.0",
    # PostgreSQL
    "psycopg2-binary==2.9.10",
    # YAML generation
    "jinja2==3.1.6",
    "pyyaml==6.0.2",
    # HTTP client
    "requests==2.32.5",
    "huggingface_hub>=0.34.4",
    "transformers>=4.55.4",
    "llm-optimizer @ git+https://github.com/bentoml/llm-optimizer.git",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
ui = [
    "streamlit==1.39.0",
    "plotly>=6.3.0",
    "matplotlib>=3.10.5",
]
cluster = [
    "kubernetes>=29.0",
]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "ruff==0.8.4",
    "mypy>=1.13",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.32",
]

[project.scripts]
neuralnav-plan = "planner.cli.planner_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/planner"]

[tool.ruff]
# Set the maximum line length
line-length = 100

# Target Python 3.11+
target-version = "py311"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.egg-info",
    "build",
    "dist",
    "generated_configs",
    "logs",
]

[tool.ruff.lint]
# Enable recommended rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modern Python syntax)
    "B",   # flake8-bugbear (common bugs)
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules that may be too strict
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults (common in FastAPI)
    "N802",   # Function name should be lowercase (some exceptions for API routes)
    "SIM108", # Use ternary operator (sometimes reduces readability)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Like Black, automatically detect the appropriate line ending
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
# Ignore import sorting in __init__.py files
"__init__.py" = ["F401", "I"]

# Test files can have longer lines and unused imports
"tests/**/*.py" = ["E501", "F401"]
"test_*.py" = ["E501", "F401"]

[tool.ruff.lint.isort]
# Organize imports into sections
known-first-party = ["planner"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "database: marks tests that require a running PostgreSQL database with benchmark data",
    "integration: marks tests as integration tests (requires Ollama and database)",
    "unit: marks tests as unit tests (no external dependencies)",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
